Rename Dialogs to Overlays to generalize to secondary ui component
